+++ meteor-app/imports/ui/charges/index.js
@@ -1,0 +1,3 @@
+import Main from './lister'
+
+export default Main