+++ scripts/modules.js
@@ -1,0 +1,15 @@
+// modules.js
+// List of generated modules
+const modules = {
+  events: "Events",
+  promos: "Promos",
+  products: "Products",
+  sessions: "Sessions",
+  parts: "Parts",
+  "service-items": "ServiceItems",
+  charges: "Charges",
+  settings: "Settings"
+};
+module.exports = modules;
+
+// End of generated modules