+++ templates/imports/ui/generated-routes.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Route } from 'react-router-dom'
+// Generated code to import all of the DB UI components
+<%= routeImports %>
+
+//
+// This file contains a list of routes for database admin pages
+// It is generated from a list of modules
+//
+export default GeneratedRoute = () => (
+  <>
+    <%= routeMarkup %>
+  </>
+)